Teacher_Student_Moral/ethics_iterative_prompt_rewriting_experiment/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ethics-iterative-prompt-rewriting-experiment"
version = "0.1.0"
description = "Rigorous prompt-rewriting experiment for ETHICS commonsense moral judgment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "accelerate>=0.30.1",
  "datasets>=2.19.1,<3.0.0",
  "google-genai>=1.13.0",
  "huggingface-hub>=0.23.2,<2.0.0",
  "matplotlib>=3.8.4,<4.0.0",
  "mlx-lm>=0.31.2",
  "numpy>=1.26.4,<3.0.0",
  "openai>=1.30.5",
  "pandas>=2.2.2,<3.0.0",
  "pydantic>=2.7.1",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
  "scikit-learn>=1.5.0,<2.0.0",
  "scipy>=1.13.1,<2.0.0",
  "tenacity>=8.3.0",
  "torch>=2.3.0,<3.0.0",
  "tqdm>=4.66.4",
  "transformers>=4.41.2,<6.0.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10.0",
  "pandas-stubs>=2.2.2.240603",
  "pytest>=8.2.1",
  "ruff>=0.5.0",
  "types-PyYAML>=6.0.12.20240311",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
warn_unreachable = true
disallow_any_generics = true

[[tool.mypy.overrides]]
module = [
  "datasets",
  "scipy",
  "scipy.*",
  "sklearn.*",
]
ignore_missing_imports = true
